requirements.txt
allure-pytest==2.10.0
allure_python_commons==2.10.0
msedge_selenium_tools==3.141.4
pandas==1.4.4
pytest==7.1.2
pytest-failed-screenshot==1.0.2
pytest-rerunfailures==10.2
pytest_steps==1.8.0
requests==2.28.1
selenium==3.141
selenium_wire==4.6.5
webdriver_manager==3.8.3
